==> 03_preprocessing.ipynb <==
##### PREPROCESSING #####

print('Starting preprocessing...')

### import libraries
import pandas as pd
import numpy as np
from datetime import datetime,date

start_time = datetime.now()
print('Start time: ', start_time)

#### SELECT INPUT AND OUTPUT FILES

input_file = '3_day_sample_aggregated.tsv.gz'
output_file = '3_day_sample_preprocessed.tsv.gz'
#input_file = '6_week_sample_aggregated.tsv.gz'
#output_file = '6_week_sample_preprocessed.tsv.gz'
#input_file = '12_week_sample_aggregated.tsv.gz'
#output_file = '12_week_sample_preprocessed.tsv.gz'
#input_file = '25_week_sample_aggregated.tsv.gz'
#output_file = '25_week_sample_preprocessed.tsv.gz'

print('Input file selected: ', input_file)
print('Output file selected', output_file)

##### LOAD DATA
print('Loading data...')

df = pd.read_csv('../data/processed_data/'+input_file, compression='gzip', sep='\t', low_memory=False, encoding='iso-8859-1', parse_dates=['hit_time_gmt', 'last_hit_time_gmt_visit', 'date_time'])

print('Loading data complete.')

### ENCODE TARGET
print('Encoding target...')

# binary encode target to be either 1 or 0
df['purchase'] = df['purchase'].apply(lambda x: 1 if x >= 1 else 0)

print('Encoding target complete.')

### ENCODE CATEGORICAL FEATURES
print('Encoding categorical features...')

categorical_cols = ['connection_type',
                    'browser',
                    'operating_system_generalized',
                    'search_engine_generalized',
                    'marketing_channel',
                    'referrer_type']

dummies = pd.get_dummies(df.loc[:, df.columns.isin(categorical_cols)], drop_first=True)
df.drop(categorical_cols, axis=1, inplace=True)
df = pd.concat([df, dummies], axis=1)

### WRITE DATAFRAME TO FILE
print('Writing dataframe to file...')

df.to_csv('../data/processed_data/'+output_file, compression='gzip', sep='\t', encoding='iso-8859-1', index=False)

print('Preprocessing complete.')
print('Run time: ', datetime.now() - start_time)
